pyproject.toml
[tool.poetry]
name = "hive metadata explorer"
version = "0.1.0"
description = "This project combines beekeeping, data management, and computer vision to turn hive photos into a structured, queryable knowledge base."
authors = ["Barbara <dagny099@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
streamlit = ">=1.15.0"
pillow = ">=9.0.0"
colorthief = ">=0.2.1"
pandas = ">=1.3.0"
requests = ">=2.27.0"
google-cloud-vision = ">=3.1.0"
streamlit-calendar = "^1.2.1"
plotly = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
